various fixes for merging the master branch
* fix: the incorrect use of clz on unsigned long data type
* ref: add 64bits variances of clzl, llog2. provide handy definition
       of most significant bit index of unsigned long and int.
* ref: uncaptialize the ILOG2
* fix: the kernel stack creation and copy algorithm, make it
  less hacky
* fix: assertion on thread state in pwake is too strong. if a
  thread is already awakened by other event or killed, just remove
  it from wait list
* fix: change the tag calculation in ext2::bTLB, so it can contain
       validity bit
* fix: a block buffer should ref again when inserting into bTLB
* fix: mixed use of size_t and unsigned int, this is not working
       on 64bits platform
* fix: boot disk prober automatically use last device it scaned
       when no blkdev can be identified as boot disk
* feat: qemu.py (autoqemu) will now skip the block device if the
        image is not found, so it provides a less disastrous
        fallback behavior
* fix: always vzalloc for the bltb buffer.

+++ lunaix-os/kernel/bcache.c
@@ -53,7 +53,7 @@ bcache_create_zone(char* name)
 
 void
 bcache_init_zone(struct bcache* cache, bcache_zone_t lru, unsigned int log_ways, 
-                   int cap, size_t blk_size, struct bcache_ops* ops)
+                   int cap, unsigned int blk_size, struct bcache_ops* ops)
 {
     // TODO handle cap
